# Minor matcher builder allocation optimization