Before:
```
 Method |     Mean |    Error |   StdDev |  Op/s |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
------- |---------:|---------:|---------:|------:|---------:|---------:|---------:|----------:|
    Dfa | 243.7 ms | 4.873 ms | 7.441 ms | 4.103 | 500.0000 | 375.0000 | 250.0000 |  88.25 MB |
```

After:
```
 Method |     Mean |    Error |   StdDev |  Op/s |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
------- |---------:|---------:|---------:|------:|---------:|---------:|---------:|----------:|
    Dfa | 275.8 ms | 5.423 ms | 5.803 ms | 3.626 | 750.0000 | 750.0000 | 125.0000 |  85.99 MB |
```